Added client.go
